Refactoring the controllers.
- Renaming the cache-headers middleware so the file matches the naming convention.
