+++ app/middleware/index.js
@@ -1,4 +1,4 @@
-const cacheHeaders = require('./cacheHeaders');
+const cacheHeaders = require('./cache-headers');
 const errorHandler = require('./error-handler');
 const i18n = require('./i18n');
